Refactor naming conventions
